# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/loan-offer-interests": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all loan offer interest with filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer line",
                    "admin"
                ],
                "summary": "Get all loan offer interest with filter",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pageNumber",
                        "name": "page[number]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pageSize",
                        "name": "page[size]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ResponseWithPaging-array_financing-offer_internal_core_entity_LoanPackageOfferInterest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/loan-offer-interests/{id}/assign-loan-contract": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create assigned loan offer interest loan contract",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer line",
                    "admin"
                ],
                "summary": "Create assigned loan offer interest loan contract",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_core_loanofferinterest_http.CreateAssignedLoanOfferInterestLoanContractRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_LoanContract"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/my-loan-offer-interests/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Investor confirm multiple loan package offer interest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer line",
                    "investor"
                ],
                "summary": "Investor confirm multiple loan package offer interest",
                "parameters": [
                    {
                        "description": "body",
                        "name": "loanPackageOfferInterestIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_core_loanofferinterest_http.InvestorConfirmLoanPackageOfferInterestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/my-loan-offer-interests/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Investor cancel loan package offer interest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer line",
                    "investor"
                ],
                "summary": "Investor cancel loan package offer interest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/my-loan-offer-interests/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Investor confirm loan package offer interest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan package offer line",
                    "investor"
                ],
                "summary": "Investor confirm loan package offer interest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.BaseResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/financing-offer_internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "financing-offer_internal_core.PagingMetaData": {
            "type": "object",
            "properties": {
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "financing-offer_internal_core_entity.AssetType": {
            "type": "string",
            "enum": [
                "UNDERLYING",
                "DERIVATIVE"
            ],
            "x-enum-varnames": [
                "AssetTypeUnderlying",
                "AssetTypeDerivative"
            ]
        },
...
        "financing-offer_internal_core_entity.CancelledReason": {
            "type": "string",
            "enum": [
                "UNKNOWN",
                "EXPIRED",
                "INVESTOR",
                "ADMIN",
                "ALTERNATIVE_OPTION",
                "HIGH_LOAN_RATE"
            ],
            "x-enum-varnames": [
                "LoanPackageOfferCancelledReasonUnknown",
                "LoanPackageOfferCancelledReasonExpired",
                "LoanPackageOfferCancelledReasonInvestor",
                "LoanPackageOfferCancelledReasonAdmin",
                "LoanPackageOfferCancelledReasonAlternativeOption",
                "LoanPackageOfferCancelledReasonHighLoanRate"
            ]
        },
...
        "financing-offer_internal_core_entity.FinancialProductLoanPackage": {
            "type": "object",
            "properties": {
                "allowEarlyPayment": {
                    "type": "boolean"
                },
                "allowExtendLoanTerm": {
                    "type": "boolean"
                },
                "brokerFirmBuyingFeeRate": {
                    "type": "number"
                },
                "brokerFirmSellingFeeRate": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "initialRate": {
                    "type": "number"
                },
                "initialRateForWithdraw": {
                    "type": "number"
                },
                "interestRate": {
                    "type": "number"
                },
                "liquidRate": {
                    "type": "number"
                },
                "loanBasketId": {
                    "type": "integer"
                },
                "loanType": {
                    "type": "string"
                },
                "maintenanceRate": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "preferentialInterestRate": {
                    "type": "number"
                },
                "preferentialPeriod": {
                    "type": "integer"
                },
                "term": {
                    "type": "integer"
                },
                "transferFee": {
                    "type": "number"
                }
            }
        },
        "financing-offer_internal_core_entity.FlowType": {
            "type": "string",
            "enum": [
                "001",
                "002",
                "ONLINE",
                "OFFLINE"
            ],
            "x-enum-varnames": [
                "FLowTypeDnseOffline",
                "FlowTypeDnseOnline",
                "Online",
                "Offline"
            ]
        },
        "financing-offer_internal_core_entity.Investor": {
            "type": "object",
            "properties": {
                "custodyCode": {
                    "type": "string"
                },
                "investorId": {
                    "type": "string"
                }
            }
        },
...
        "financing-offer_internal_core_entity.LoanContract": {
            "type": "object",
            "properties": {
                "accountNo": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "guaranteedEndAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "investorId": {
                    "type": "string"
                },
                "loanId": {
                    "type": "integer"
                },
                "loanInterestId": {
                    "type": "integer"
                },
                "loanPackageAccountId": {
                    "type": "integer"
                },
                "loanProductIdRef": {
                    "type": "integer"
                },
                "symbolId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.LoanPackageOffer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "expiredAt": {
                    "type": "string"
                },
                "flowType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.FlowType"
                },
                "id": {
                    "type": "integer"
                },
                "loanPackageOfferInterests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOfferInterest"
                    }
                },
                "loanPackageRequest": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequest"
                },
                "loanPackageRequestId": {
                    "type": "integer"
                },
                "offeredBy": {
                    "type": "string"
                },
                "symbol": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Symbol"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.LoanPackageOfferInterest": {
            "type": "object",
            "properties": {
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "cancelledAt": {
                    "type": "string"
                },
                "cancelledBy": {
                    "type": "string"
                },
                "cancelledReason": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.CancelledReason"
                },
                "contractSize": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "feeRate": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "initialRate": {
                    "type": "number"
                },
                "interestRate": {
                    "type": "number"
                },
                "limitAmount": {
                    "type": "number"
                },
                "loanContract": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanContract"
                },
                "loanId": {
                    "type": "integer"
                },
                "loanPackageOffer": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOffer"
                },
                "loanPackageOfferId": {
                    "type": "integer"
                },
                "loanRate": {
                    "type": "number"
                },
                "scoreGroupInterestId": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOfferInterestStatus"
                },
                "submissionSheetDetailId": {
                    "type": "integer"
                },
                "term": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.LoanPackageOfferInterestStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "CANCELLED",
                "SIGNED",
                "PACKAGE_CREATED",
                "PACKAGE_CREATING"
            ],
            "x-enum-varnames": [
                "LoanPackageOfferInterestStatusPending",
                "LoanPackageOfferInterestStatusCancelled",
                "LoanPackageOfferInterestStatusSigned",
                "LoanPackageOfferInterestStatusLoanPackageCreated",
                "LoanPackageOfferInterestStatusCreatingLoanPackage"
            ]
        },
        "financing-offer_internal_core_entity.LoanPackageRequest": {
            "type": "object",
            "properties": {
                "accountNo": {
                    "type": "string"
                },
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "contractSize": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "guaranteedDuration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "initialRate": {
                    "type": "number"
                },
                "investor": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.Investor"
                },
                "investorId": {
                    "type": "string"
                },
                "limitAmount": {
                    "type": "number"
                },
                "loanRate": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequestStatus"
                },
                "symbolId": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageRequestType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.LoanPackageRequestStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "CONFIRMED"
            ],
            "x-enum-varnames": [
                "LoanPackageRequestStatusPending",
                "LoanPackageRequestStatusConfirmed"
            ]
        },
        "financing-offer_internal_core_entity.LoanPackageRequestType": {
            "type": "string",
            "enum": [
                "FLEXIBLE",
                "GUARANTEED"
            ],
            "x-enum-varnames": [
                "LoanPackageRequestTypeFlexible",
                "LoanPackageRequestTypeGuaranteed"
            ]
        },
...
        "financing-offer_internal_core_entity.Symbol": {
            "type": "object",
            "properties": {
                "assetType": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.AssetType"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastUpdatedBy": {
                    "type": "string"
                },
                "scores": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScore"
                    }
                },
                "status": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolStatus"
                },
                "stockExchangeId": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.SymbolScore": {
            "type": "object",
            "properties": {
                "affectedFrom": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "status": {
                    "enum": [
                        "ACTIVE",
                        "INACTIVE"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScoreStatus"
                        }
                    ]
                },
                "symbolId": {
                    "type": "integer"
                },
                "type": {
                    "enum": [
                        "MANUAL",
                        "SYSTEM"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/financing-offer_internal_core_entity.SymbolScoreType"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_core_entity.SymbolScoreStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "SymbolScoreStatusActive",
                "SymbolScoreStatusInactive"
            ]
        },
        "financing-offer_internal_core_entity.SymbolScoreType": {
            "type": "string",
            "enum": [
                "MANUAL",
                "SYSTEM"
            ],
            "x-enum-varnames": [
                "SymbolScoreTypeManual",
                "SymbolScoreTypeSystem"
            ]
        },
        "financing-offer_internal_core_entity.SymbolStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "SymbolStatusActive",
                "SymbolStatusInactive"
            ]
        },
...
        "financing-offer_internal_handler.BaseResponse-financing-offer_internal_core_entity_LoanContract": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.LoanContract"
                }
            }
        },
...
        "financing-offer_internal_handler.BaseResponse-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "financing-offer_internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "financing-offer_internal_handler.ResponseWithPaging-array_financing-offer_internal_core_entity_LoanPackageOfferInterest": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/financing-offer_internal_core_entity.LoanPackageOfferInterest"
                    }
                },
                "metaData": {
                    "$ref": "#/definitions/financing-offer_internal_core.PagingMetaData"
                }
            }
        },
...
        "internal_core_loanofferinterest_http.CreateAssignedLoanOfferInterestLoanContractRequest": {
            "type": "object",
            "properties": {
                "loanPackage": {
                    "$ref": "#/definitions/financing-offer_internal_core_entity.FinancialProductLoanPackage"
                },
                "loanPackageAccountId": {
                    "type": "integer"
                },
                "loanProductIdRef": {
                    "type": "integer"
                }
            }
        },
        "internal_core_loanofferinterest_http.InvestorConfirmLoanPackageOfferInterestRequest": {
            "type": "object",
            "required": [
                "loanPackageOfferInterestIds"
            ],
            "properties": {
                "loanPackageOfferInterestIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}